refactor: move framework code into file with type safety
This commit moves the framework code into the TS compilation with type
safety, adding types where necessary, and then leveraging esbuild to
construct a single file ESM bundle that can be included in the wrapper
HTML.
