+++ package.json
@@ -17,7 +17,7 @@
     "setup": "curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.1/install.sh | bash && unset npm_config_prefix && . ~/.nvm/nvm.sh && nvm install && nvm use && npm i && npx playwright install chromium --with-deps chromium",
     "clean": "del ./dist",
     "test": "web-test-runner --coverage",
-    "build": "tsc --project ./tsconfig.lib.json && cpy ./package.json dist/lib/ && cpy ./README.md dist/lib/",
+    "build": "bash build.sh",
     "publish": "cd ./dist/lib && npm publish --access public"
   },
   "repository": {
@@ -32,8 +32,7 @@
   "homepage": "https://github.com/blueprintui/web-test-runner-jasmine",
   "dependencies": {
     "@web/test-runner": "^0.19.0",
-    "@web/test-runner-core": "^0.13.4",
-    "jasmine": "^5.5.0"
+    "@web/test-runner-core": "^0.13.4"
   },
   "devDependencies": {
     "@types/jasmine": "5.1.5",
@@ -41,6 +40,9 @@
     "@web/test-runner-playwright": "0.11.0",
     "cpy-cli": "5.0.0",
     "del-cli": "6.0.0",
+    "esbuild": "^0.25.2",
+    "jasmine": "^5.5.0",
+    "jasmine-core": "^5.6.0",
     "playwright": "1.49.1",
     "typescript": "5.7.3"
   }
