+++ src/jasmine-types.d.ts
@@ -1,0 +1,9 @@
+declare namespace jasmine {
+  interface SpecResult {
+    parentSuiteId: string;
+  }
+
+  interface SuiteResult {
+    parentSuiteId: string;
+  }
+}
